+++ Forge/src/main/kotlin/gay/object/hexdebug/forge/interop/ForgeHexDebugEmiPlugin.kt
@@ -6,7 +6,7 @@ import dev.emi.emi.api.EmiRegistry
 import gay.`object`.hexdebug.interop.HexDebugEmiPlugin
 
 @EmiEntrypoint
-object ForgeHexDebugEmiPlugin : EmiPlugin {
+class ForgeHexDebugEmiPlugin : EmiPlugin {
     override fun register(registry: EmiRegistry) {
         HexDebugEmiPlugin.register(registry)
     }